requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-genai
motor  
pymongo
python-multipart  
reportlab 
passlib==1.7.4
bcrypt==3.2.2
python-jose[cryptography]
